Clean up benchmark code
